Moved Auth and UrlKey related functions into logic/security.go
Changed usages accordingly
